Fixed libGLESv2 conversion warnings
There were only a handful of integer conversion warnings in
libGLESv2, so I fixed them rather than silencing them.

Reviewed-on: https://swiftshader-review.googlesource.com/7132
Tested-by: Alexis Hétu <sugoi@google.com>

+++ src/OpenGL/libGLESv2/IndexDataManager.cpp
@@ -121,7 +121,7 @@ GLenum IndexDataManager::prepareIndexData(GLenum type, GLuint start, GLuint end,
 		computeRange(type, indices, count, &translated->minIndex, &translated->maxIndex);
 
 		translated->indexBuffer = staticBuffer;
-		translated->indexOffset = offset;
+		translated->indexOffset = static_cast<unsigned int>(offset);
 	}
 	else
 	{
